Move `get-api-url` to `config`
Also: update tests for modern syntax
